feat: centralise logs
