+++ build_pipegene/scripts/github_actions.py
@@ -1,7 +1,8 @@
 import click
 
 from validations import validate_pipeline
-from pipeline_parameters import PipelineParametersHandler
+from scripts.utils.pipeline_parameters import PipelineParametersHandler
+from scripts.utils.log_pipe_params import log_pipeline_params
 
 @click.group()
 def cli():
@@ -10,7 +11,7 @@ def cli():
 @cli.command("validate_pipeline")
 def validate_pipeline_command():
     handler = PipelineParametersHandler()
-    handler.log_params()
+    log_pipeline_params(handler.params.copy())
     validate_pipeline(handler.params)
 
 if __name__ == "__main__":
