+++ build_pipegene/scripts/main.py
@@ -2,7 +2,8 @@ import click
 
 from gitlab_ci import build_pipeline
 from validations import validate_pipeline
-from pipeline_parameters import PipelineParametersHandler
+from scripts.utils.pipeline_parameters import PipelineParametersHandler
+from scripts.utils.log_pipe_params import log_pipeline_params
 
 @click.group(chain=True)
 def gcip():
@@ -14,7 +15,7 @@ def generate_pipeline():
 
 def perform_generation():
     handler = PipelineParametersHandler()
-    handler.log_params()
+    log_pipeline_params(handler.params.copy())
     validate_pipeline(handler.params)
     build_pipeline(handler.params)
